// apps/admin-shadcn/src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        ArticleResponseDto: {
            /** @example 550e8400-e29b-41d4-a716-446655440000 */
            id: string;
            /** @example Getting Started with NestJS */
            title: string;
            /** @example NestJS is a progressive Node.js framework... */
            content: string;
            /** @example getting-started-with-nestjs */
            slug: string;
            /**
             * @example draft
             * @enum {string}
             */
            status: "draft" | "published" | "archived";
            /**
             * @example [
             *       "nestjs",
             *       "typescript"
             *     ]
             */
            tags: string[];
            /**
             * @example tech
             * @enum {string}
             */
            category: "tech" | "design" | "product" | "other";
            /** @example John Doe */
            author: string;
            /** @example 42 */
            viewCount: number;
            /** @example false */
            isPinned: boolean;
            /**
             * Format: date-time
             * @example 2026-01-01T00:00:00.000Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @example 2026-01-01T00:00:00.000Z
             */
            updatedAt: string;
            /** @example 2026-01-01T00:00:00.000Z */
            publishedAt: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}